Freezing issue fixwork includes:
1.	Changed a return to break in check for water tiles - State.as
2.	Now removing event listener for takebreath.
3.	Reinstated the Preloader

Actionscript Settings:

	1.	Set jpg quality to 50
	2.	Disabled hidden layers
	3.	Disabled Generate Size Report

When searching for location of this problem; I put some debug traces in State.as and Player.as. I found that both were freezing after a super.update() to the class which tells me the freeze is located elsewhere because the update function was completing each time and hanging in between.

This issue is not completely resolved only improved slightly from previous push. If you need to disable preloader removal again go ahead but check first because it did not mess the Menu up for me. We need to continue focusing on this problem and fixing it before adding much more code.